Weekly Software Update
* fixed flickering bug in raycasting
* improved vct with wieght function
* added cosine scaling to raycasting

+++ vg/vg/ambientocclusion.hpp
@@ -30,7 +30,7 @@ class ambientocclusion
 		GLuint normal_texture;
 		GLuint occlusion_texture;
 		
-		unsigned int samples;
+		float samples;
 
 		GLuint framebuffer;
 		GLuint renderbuffer;
